# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Orchestration backend for the automated short-form video pipeline, plus the original book sample endpoints.",
        "title": "Pom Pom Hollow Production API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pipeline/control": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "Show what is paused",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.PipelineControl"
                            }
                        }
                    }
                }
            }
        },
        "/pipeline/control/pause": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "The operator's stop switch. Workers can no longer take work, so the pipeline drains rather than halting mid-job: whatever is already claimed finishes, nothing new starts. Pass a stage as the scope to pause one step — pausing SCHEDULED holds uploads while the rest keeps producing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "Stop production",
                "parameters": [
                    {
                        "description": "What to pause and why",
                        "name": "pause",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PauseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PipelineControl"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pipeline/control/resume": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lifts a pause. Workers begin taking work at their next poll; nothing needs restarting.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "Start production again",
                "parameters": [
                    {
                        "description": "What to resume",
                        "name": "pause",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PauseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PipelineControl"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/episodes/{id}/fail": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Called by a worker that has given up. The stage is retried until the attempt limit, after which the episode is parked in FAILED for a person to look at rather than being retried forever.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Report that a job could not be finished",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "What went wrong",
                        "name": "failure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.FailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/episodes/{id}/retry": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "The one sanctioned way out of FAILED. Restores the stage the episode failed at and clears the attempt count, so a fix to the underlying problem can be tried without recreating the episode.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "Put a failed episode back to work",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/overview": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Everything one operator needs to see at a glance: whether production is running, how much sits at each stage, what is waiting on a human, what has failed, and what has gone quiet.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "The whole operation on one screen",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PipelineOverview"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Episode": {
            "type": "object",
            "properties": {
                "ai_disclosure": {
                    "type": "boolean"
                },
                "attempts": {
                    "description": "Attempts counts how many times the current stage has been tried. It\nresets whenever the episode moves on, so it measures this stage rather\nthan the episode's whole history.",
                    "type": "integer"
                },
                "claimed_at": {
                    "type": "string"
                },
                "claimed_by": {
                    "description": "ClaimedBy is the worker currently holding this episode, and ClaimedAt\nis when it took it. The claim is a lease: it expires so a worker that\ndies mid-job does not strand the episode. Both are cleared whenever the\nepisode changes stage.",
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "curriculum_topic_id": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "failed_from": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "id": {
                    "type": "integer"
                },
                "last_error": {
                    "type": "string"
                },
                "made_for_kids": {
                    "description": "MadeForKids drives the YouTube selfDeclaredMadeForKids flag. It defaults\nto true because every episode of this show is children's content, and\ngetting this wrong is a COPPA problem rather than a cosmetic one.",
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "script_text": {
                    "type": "string"
                },
                "song_lyrics": {
                    "type": "string"
                },
                "song_title": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "synopsis": {
                    "type": "string"
                },
                "tiktok_video_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "youtube_video_id": {
                    "type": "string"
                }
            }
        },
...
        "main.EpisodeStatus": {
            "type": "string",
            "enum": [
                "IDEA_BACKLOG",
                "SCRIPT_DRAFT",
                "SCRIPT_APPROVED",
                "VO_GENERATED",
                "MUSIC_GENERATED",
                "ANIMATION_RENDERED",
                "ASSEMBLED",
                "QA_REVIEW",
                "PLATFORM_ADAPTED",
                "SCHEDULED",
                "PUBLISHED",
                "ANALYZED",
                "CANCELLED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusIdeaBacklog",
                "StatusScriptDraft",
                "StatusScriptApproved",
                "StatusVOGenerated",
                "StatusMusicGenerated",
                "StatusAnimationRendered",
                "StatusAssembled",
                "StatusQAReview",
                "StatusPlatformAdapted",
                "StatusScheduled",
                "StatusPublished",
                "StatusAnalyzed",
                "StatusCancelled",
                "StatusFailed"
            ]
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.FailRequest": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
        "main.PauseRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                }
            }
        },
        "main.PipelineControl": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "paused": {
                    "type": "boolean"
                },
                "paused_at": {
                    "type": "string"
                },
                "paused_by": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "main.PipelineOverview": {
            "type": "object",
            "properties": {
                "awaiting_review": {
                    "type": "integer"
                },
                "claimed_now": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "in_flight": {
                    "type": "integer"
                },
                "open_alerts": {
                    "type": "integer"
                },
                "pause_reason": {
                    "type": "string"
                },
                "paused": {
                    "type": "boolean"
                },
                "paused_stages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "published_total": {
                    "type": "integer"
                },
                "stages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.StageCount"
                    }
                },
                "stuck": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Episode"
                    }
                }
            }
        },
        "main.StageCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "paused": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}